+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Run acceptance tests with latest version 9.8.0
+
+### Fixed
+- Fix setting PID file location on SonarQube 9.6 and later ([#24])
+
 ## [5.0.0] - 2022-10-31
 This is a new major release and contains breaking changes. Of note is the
 new requirement to specify the `$version` parameter using SonarQubes
@@ -135,6 +143,7 @@ migrate from maestrodev/sonarqube to this version with only minor modifications.
 [4.0.0]: https://github.com/markt-de/puppet-sonarqube/compare/v3.1.0...v4.0.0
 [3.1.0]: https://github.com/markt-de/puppet-sonarqube/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/markt-de/puppet-sonarqube/compare/v2.6.7...v3.0.0
+[#24]: https://github.com/markt-de/puppet-sonarqube/pull/24
 [#23]: https://github.com/markt-de/puppet-sonarqube/pull/23
 [#22]: https://github.com/markt-de/puppet-sonarqube/pull/22
 [#21]: https://github.com/markt-de/puppet-sonarqube/pull/21